hrtimer: Don't apply the BUG_ON() in migrate_hrtimer_list() for isolation
commit 25c3b4aa9277 ("hrtimer: Don't drop the base lock when migration
during isolation") added the BUG_ON() in migrate_hrtimer_list() for
CPU isolation as well. This BUG_ON() can hit in the isolation path
in the following scenario.

- CPU#2 is isolating CPU#0. CPU#0 is marked as isolated but hrtimer
migration is not yet started.

- A hrtimer callback is in progress on CPU#0.

- CPU#1 queue the same hrtimer while the callback is running on
CPU#0. As CPU#0 is isolated, we try to queue the hrtimer on a
different CPU, but switch_hrtimer_base() does not allow it as
the callback is in progress. The hrtimer is queued on CPU#0.

- CPU#2 acquires CPU#0's base lock and try to migrate the hrtimer.
As it's callback in progress, we hit the BUG_ON().

Change-Id: I745856a62e8065644edd551d2a5f54f0ed912f62
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>